backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.9.0
httpx>=0.27.0
python-dotenv>=1.0.1
llama-cpp-python>=0.3.0
